requirements.txt
yfinance>=0.2.18
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0